Set dashboard as home and load Woo products from DB
