+++ src/app/api/woo/products/route.ts
@@ -1,5 +1,8 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { getShopConfig, fetchWooProducts } from '@/lib/wooApi';
+import { fetchWooProducts } from '@/lib/wooApi';
+import { db } from '@/lib/db';
+import { shops } from '@/lib/schema';
+import { eq } from 'drizzle-orm';
 
 export async function GET(req: NextRequest) {
   const { searchParams } = new URL(req.url);
@@ -7,12 +10,18 @@ export async function GET(req: NextRequest) {
   if (!shopId) {
     return NextResponse.json({ error: 'Missing shopId' }, { status: 400 });
   }
-  const shop = getShopConfig(shopId);
-  if (!shop) {
+  const [shop] = await db.select().from(shops).where(eq(shops.id, shopId));
+  if (!shop || !shop.apiKey || !shop.apiSecret) {
     return NextResponse.json({ error: 'Shop not found' }, { status: 404 });
   }
   try {
-    const products = await fetchWooProducts(shop);
+    const products = await fetchWooProducts({
+      id: shop.id,
+      name: shop.name,
+      url: shop.url,
+      consumer_key: shop.apiKey,
+      consumer_secret: shop.apiSecret,
+    });
     return NextResponse.json(products);
   } catch (err: unknown) {
     const message = err instanceof Error ? err.message : 'Unknown error';
